Create dashcam.py
This script accesses a cam connected to raspberry pi (or any other machine) and creates 10 second videos. The name of the video is the time stamp. 
It saves each video in a given folder and starts a new one. 
This can be used as a dashcam loop recording. A couple of things are missing so far:
- Need to update the boot up file for raspian in order to start recording every time the pi boots up.
- Also need to delete old videos in order to create new ones else memory gets full in a while.